+++ app/Http/Controllers/ApiOfficesController.php
@@ -24,15 +24,28 @@ class ApiOfficesController extends Controller
         return response()->json($office, 201);
     }
 
-    public function update(Request $request, Office $office)
+    public function update(Request $request, $id)
     {
+        $office = Office::find($id);
+        if (!$office) {
+            return response()->json(array(
+                'error' => array(
+                    'message' => 'Bad request. The standard option for requests that fail to pass validation.'
+                )
+            ), 400);
+        }
         $office->update($request->all());
         return response()->json($office, 200);
     }
 
-    public function delete(Office $office)
-    {
-        $office->delete();
+    public function delete(Request $request, $id) {
+        $office = Office::find($id);
+
+        if ($office) {
+            $office->delete();
+        }
+
         return response()->json(null, 204);
     }
+
 }
